requirements.txt
pandas
numpy
scikit-learn
xgboost
shap
mlflow
fastapi
uvicorn[standard]
pydantic
sqlalchemy
psycopg2-binary
python-dotenv
streamlit
plotly
dbt-postgres
ruff
black
mypy
pytest
pre-commit
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
psycopg2-binary
joblib
pandas
scikit-learn